auth_service/requirements.txt
fastapi
python-dotenv
psycopg2-binary
python-jose[cryptography]
requests
paho-mqtt
uvicorn
PyJWT